# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProviEmplea API",
        "description": "API REST para la plataforma de empleo ProviEmplea de Providencia. Permite gestionar talentos, empresas y procesos de selección con curriculum ciego.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/personas": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Listar personas (CV ciego)",
                "description": "Obtiene talentos activos en formato de CV ciego (sin datos personales identificables). Rate limit: 120 solicitudes por minuto por IP. Se recomienda aplicar filtros para reducir payload y paginar si el volumen crece.",
                "operationId": "getPersonas",
                "parameters": [
                    {
                        "name": "validado",
                        "in": "query",
                        "description": "Filtrar por validación",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "nivel_educacional",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "basica",
                                "media",
                                "tecnica",
                                "universitaria",
                                "postgrado"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonaListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Personas"
                ],
                "summary": "Registrar nueva persona/talento",
                "description": "Crea un perfil de talento. El código se genera automáticamente.",
                "operationId": "createPersona",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Persona creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonaResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/personas/{persona}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener persona por ID",
                "description": "Consulta de lectura con rate limit de 120 solicitudes por minuto por IP.",
                "operationId": "getPersona",
                "parameters": [
                    {
                        "name": "persona",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonaResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Personas"
                ],
                "summary": "Actualizar persona",
                "operationId": "updatePersona",
                "parameters": [
                    {
                        "name": "persona",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Persona actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonaResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Personas"
                ],
                "summary": "Desactivar persona",
                "description": "Desactiva el perfil sin eliminarlo de la base de datos. Rate limit: 30 solicitudes por minuto por IP.",
                "operationId": "deletePersona",
                "parameters": [
                    {
                        "name": "persona",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona desactivada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/personas/{persona}/validar": {
            "patch": {
                "tags": [
                    "Personas"
                ],
                "summary": "Validar persona (solo administración)",
                "description": "Marca a una persona como validada para que aparezca en la vitrina. Rate limit: 30 solicitudes por minuto por IP.",
                "operationId": "validarPersona",
                "parameters": [
                    {
                        "name": "persona",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona validada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonaValidationResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PersonaInput": {
                "title": "Persona Input",
                "description": "Datos para crear o actualizar una persona",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@example.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+56912345678",
                        "nullable": true
                    },
                    "resumen": {
                        "type": "string",
                        "nullable": true
                    },
                    "nivel_educacional": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "basica",
                            "media",
                            "tecnica",
                            "universitaria",
                            "postgrado"
                        ]
                    },
                    "titulo_carrera": {
                        "type": "string",
                        "nullable": true
                    },
                    "anio_egreso": {
                        "type": "integer",
                        "example": 2020,
                        "nullable": true
                    },
                    "anios_experiencia": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "areas_experiencia": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Desarrollo Web"
                        ],
                        "nullable": true
                    },
                    "competencias": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "PHP",
                            "Laravel"
                        ],
                        "nullable": true
                    },
                    "rango_renta": {
                        "type": "string",
                        "example": "500k-800k",
                        "nullable": true
                    },
                    "tipo_jornada": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "completa",
                            "part-time",
                            "por-horas"
                        ]
                    },
                    "modalidad": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "presencial",
                            "remoto",
                            "hibrido"
                        ]
                    },
                    "persona_discapacidad": {
                        "type": "boolean",
                        "example": false,
                        "nullable": true
                    },
                    "portafolio_url": {
                        "type": "string",
                        "format": "url",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PersonaCVCiego": {
                "title": "Persona CV Ciego",
                "description": "Vista pública del talento sin datos personales identificables",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "codigo_talento": {
                        "type": "string",
                        "example": "PROV-2026-A1B2"
                    },
                    "resumen": {
                        "type": "string",
                        "nullable": true
                    },
                    "nivel_educacional": {
                        "type": "string",
                        "nullable": true
                    },
                    "titulo_carrera": {
                        "type": "string",
                        "nullable": true
                    },
                    "anio_egreso": {
                        "type": "integer",
                        "nullable": true
                    },
                    "anios_experiencia": {
                        "type": "integer",
                        "example": 5
                    },
                    "tipo_jornada": {
                        "type": "string",
                        "nullable": true
                    },
                    "modalidad": {
                        "type": "string",
                        "nullable": true
                    },
                    "persona_discapacidad": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Respuesta estándar de error de la API",
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Recurso no encontrado."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "El campo email es obligatorio."
                            ]
                        },
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ActionMessage": {
                "title": "Action Message",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Persona desactivada exitosamente."
                    }
                },
                "type": "object"
            },
            "PersonaValidationResult": {
                "title": "Persona Validation Result",
                "required": [
                    "message",
                    "data"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Persona validada exitosamente."
                    },
                    "data": {
                        "$ref": "#/components/schemas/Persona"
                    }
                },
                "type": "object"
            },
...
            "PersonaListResponse": {
                "title": "Persona List Response",
                "required": [
                    "success",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PersonaCVCiego"
                        }
                    }
                },
                "type": "object"
            },
            "PersonaResponse": {
                "title": "Persona Response",
                "required": [
                    "success",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "$ref": "#/components/schemas/Persona"
                    }
                },
                "type": "object"
            },
            "PersonaValidationResponse": {
                "title": "Persona Validation Response",
                "required": [
                    "success",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "$ref": "#/components/schemas/PersonaValidationResult"
                    }
                },
                "type": "object"
            },
            "MessageResponse": {
                "title": "Message Response",
                "required": [
                    "success",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "$ref": "#/components/schemas/ActionMessage"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}